Make calendars accessible without login
